# speed up filename matches in collections module