Precomputing the downcased search string, avoiding setting the Visible status on rows where it has not changed, and other minor optimizations substantially reduce the time needed to refresh the display when typing a partial filename for a filename filtering rule.

With a library containing 331k images, typing the base name of an image with intermediate match counts of ~250k, 243k, 62k, 10k, 1000, 100, 10, and 1, the time for the gtk_tree_model_foreach(list_match_string) call goes from
  7649ms - 1493ms - 2059ms - 5248ms - 1048ms - 934ms - 920ms - 914ms
to
  6328ms - 306ms -  1329ms - 4581ms -  411ms - 322ms - 317ms - 304ms